+++ src/axom/quest/detail/MarchingCubesImpl.hpp
@@ -7,13 +7,12 @@
 
 // Implementation requires Conduit.
 #ifndef AXOM_USE_CONDUIT
-  #error "MarchingCubesHybridParallel.hpp requires conduit"
+  #error "MarchingCubesImpl.hpp requires conduit"
 #endif
 #include "conduit_blueprint.hpp"
 
 #include "axom/core/execution/execution_space.hpp"
 #include "axom/quest/ArrayIndexer.hpp"
-#include "axom/quest/detail/marching_cubes_lookup.hpp"
 #include "axom/quest/MeshViewUtil.hpp"
 #include "axom/primal/geometry/Point.hpp"
 #include "axom/primal/constants.hpp"
@@ -35,26 +34,18 @@ namespace marching_cubes
   classes MarchCubes and MarchingCubesSingleDomain.
 
   ExecSpace is the general execution space, like axom::SEQ_EXEC and
-  axom::CUDA_EXEC<256>.  SequentialExecSpace is used for loops that
-  cannot be parallelized but must access data allocated for ExecSpace.
-  Use something like axom::SEQ_EXEC or axom::CUDA_EXEC<1>.
+  axom::CUDA_EXEC<256>.
 
-  The difference between MarchingCubesHybridParallel and MarchingCubesFullParallel
-  is how they compute indices for the unstructured surface mesh.
-  - MarchingCubesHybridParallel uses sequential loop, and skips over parents
-    cells that don't touch the surface contour.  It processes less data
-    but is not data-parallel.
-  - MarchingCubesFullParallel that checks all parents cells.
-    It process more data but is data-parallel.
-  We've observed that MarchingCubesHybridParallel is faster for seq policy
-  and MarchingCubesHybridParallel is faster on the GPU.
+  See MarchingCubesImpl for the difference between that class and
+  MarchingCubesImpl.
 */
 template <int DIM, typename ExecSpace, typename SequentialExecSpace>
-class MarchingCubesHybridParallel : public MarchingCubesSingleDomain::ImplBase
+class MarchingCubesImpl : public MarchingCubesSingleDomain::ImplBase
 {
 public:
   using Point = axom::primal::Point<double, DIM>;
   using MIdx = axom::StackArray<axom::IndexType, DIM>;
+  using FacetIdType = int;
   using LoopPolicy = typename execution_space<ExecSpace>::loop_policy;
   using ReducePolicy = typename execution_space<ExecSpace>::reduce_policy;
   using SequentialLoopPolicy =
@@ -65,7 +56,6 @@ public:
     @param dom Blueprint structured mesh domain
     @param topologyName Name of mesh topology (see blueprint
            mesh documentation)
-    @param fcnFieldName Name of nodal function in dom
     @param maskFieldName Name of integer cell mask function is in dom
 
     Set up views to domain data and allocate other data to work on the
@@ -115,20 +105,11 @@ public:
     m_contourVal = contourVal;
   }
 
-#if 0
-  void computeContourMesh() override
-  {
-    markCrossings();
-    scanCrossings();
-    computeContour();
-  }
-#endif
-
   /*!
     @brief Implementation of virtual markCrossings.
 
     Virtual methods cannot be templated, so this implementation
-    delegates to a name templated on DIM.
+    delegates to an implementation templated on DIM.
   */
   void markCrossings() override { markCrossings_dim(); }
 
@@ -191,7 +172,7 @@ public:
   }
 
   /*!
-    @brief Implementation used by MarchingCubesHybridParallel::markCrossings_dim()
+    @brief Implementation used by MarchingCubesImpl::markCrossings_dim()
     containing just the objects needed for that part, to be made available
     on devices.
   */
@@ -268,15 +249,113 @@ public:
     }
   };  // MarkCrossings_Util
 
-  /*!
-    @brief Populate the 1D m_crossings array, one entry for each
-    parent cell that crosses the contour.
-
-    We sum up the number of contour surface cells from the crossings,
-    allocate space, then populate it.
-  */
   void scanCrossings() override
   {
+    constexpr MarchingCubesDataParallelism autoPolicy =
+      std::is_same<ExecSpace, axom::SEQ_EXEC>::value ? MarchingCubesDataParallelism::hybridParallel :
+      std::is_same<ExecSpace, axom::OMP_EXEC>::value ? MarchingCubesDataParallelism::hybridParallel :
+      MarchingCubesDataParallelism::fullParallel;
+
+    if(m_dataParallelism ==
+       axom::quest::MarchingCubesDataParallelism::hybridParallel ||
+       (m_dataParallelism == axom::quest::MarchingCubesDataParallelism::byPolicy &&
+        autoPolicy == MarchingCubesDataParallelism::hybridParallel))
+    {
+      scanCrossings_hybridParallel();
+    }
+    else
+    {
+      scanCrossings_fullParallel();
+    }
+  }
+
+  void scanCrossings_fullParallel()
+  {
+#if defined(AXOM_USE_RAJA)
+  #ifdef __INTEL_LLVM_COMPILER
+    // Intel oneAPI compiler segfaults with OpenMP RAJA scan
+    using ScanPolicy =
+      typename axom::execution_space<axom::SEQ_EXEC>::loop_policy;
+  #else
+    using ScanPolicy = typename axom::execution_space<ExecSpace>::loop_policy;
+  #endif
+#endif
+    const axom::IndexType parentCellCount = m_caseIds.size();
+    auto caseIdsView = m_caseIds.view();
+
+    //
+    // Initialize crossingFlags to 0 or 1, prefix-sum it, and count the
+    // crossings.
+    //
+    // All 1D array data alligns with m_caseId, which is column-major
+    // (the default for Array class), leading to *column-major* parent
+    // cell ids, regardless of the ordering of the input mesh data.
+    //
+
+    axom::Array<axom::IndexType, 1, MemorySpace> crossingFlags(parentCellCount);
+    auto crossingFlagsView = crossingFlags.view();
+    axom::for_all<ExecSpace>(
+      0,
+      parentCellCount,
+      AXOM_LAMBDA(axom::IndexType parentCellId) {
+        auto numContourCells = num_contour_cells(caseIdsView.flatIndex(parentCellId));
+        crossingFlagsView[parentCellId] = bool(numContourCells);
+      });
+
+    axom::Array<axom::IndexType, 1, MemorySpace> scannedFlags(1 + parentCellCount);
+    auto scannedFlagsView = scannedFlags.view();
+    RAJA::inclusive_scan<ScanPolicy>(
+      RAJA::make_span(crossingFlags.data(), parentCellCount),
+      RAJA::make_span(scannedFlags.data() + 1, parentCellCount),
+      RAJA::operators::plus<axom::IndexType> {});
+
+    axom::copy(&m_crossingCount,
+               scannedFlags.data() + scannedFlags.size() - 1,
+               sizeof(axom::IndexType));
+
+    //
+    // Generate crossing-cells index list and corresponding facet counts.
+    //
+
+    m_crossingParentIds.resize(m_crossingCount);
+    m_facetIncrs.resize(m_crossingCount);
+    m_firstFacetIds.resize(1 + m_crossingCount);
+
+    auto crossingParentIdsView = m_crossingParentIds.view();
+    auto facetIncrsView = m_facetIncrs.view();
+
+    auto loopBody = AXOM_LAMBDA(axom::IndexType parentCellId) {
+      if(scannedFlagsView[parentCellId] !=
+         scannedFlagsView[1+parentCellId]) {
+        auto crossingId = scannedFlagsView[parentCellId];
+        auto facetIncr = num_contour_cells(caseIdsView.flatIndex(parentCellId));
+        crossingParentIdsView[crossingId] = parentCellId;
+        facetIncrsView[crossingId] = facetIncr;
+      }
+    };
+    axom::for_all<ExecSpace>(0, parentCellCount, loopBody);
+
+    //
+    // Prefix-sum the facets counts to get first facet id for each crossing
+    // and the total number of facets.
+    //
+
+    RAJA::inclusive_scan<ScanPolicy>(
+      RAJA::make_span(m_facetIncrs.data(), m_crossingCount),
+      RAJA::make_span(m_firstFacetIds.data() + 1, m_crossingCount),
+      RAJA::operators::plus<axom::IndexType> {});
+
+    axom::copy(&m_facetCount,
+               m_firstFacetIds.data() + m_firstFacetIds.size() - 1,
+               sizeof(axom::IndexType));
+    m_firstFacetIds.resize(m_crossingCount);
+  }
+
+  void scanCrossings_hybridParallel()
+  {
+    //
+    // Compute number of crossings in m_caseIds
+    //
     const axom::IndexType parentCellCount = m_caseIds.size();
     auto caseIdsView = m_caseIds.view();
 #if defined(AXOM_USE_RAJA)
@@ -296,12 +375,15 @@ public:
     m_crossingCount = vsum;
 #endif
 
-    m_crossings.resize(m_crossingCount, {0, 0});
-    axom::ArrayView<CrossingInfo, 1, MemorySpace> crossingsView =
-      m_crossings.view();
+    //
+    // Allocate space for crossing info
+    //
+    m_crossingParentIds.resize(m_crossingCount);
+    m_facetIncrs.resize(m_crossingCount);
+    m_firstFacetIds.resize(1 + m_crossingCount);
 
-    axom::Array<int, 1, MemorySpace> addCells(m_crossingCount, m_crossingCount);
-    const axom::ArrayView<int, 1, MemorySpace> addCellsView = addCells.view();
+    auto crossingParentIdsView = m_crossingParentIds.view();
+    auto facetIncrsView = m_facetIncrs.view();
 
     axom::IndexType* crossingId = axom::allocate<axom::IndexType>(
       1,
@@ -313,18 +395,16 @@ public:
       auto ccc = num_contour_cells(caseId);
       if(ccc != 0)
       {
-        addCellsView[*crossingId] = ccc;
-        crossingsView[*crossingId].caseNum = caseId;
-        crossingsView[*crossingId].parentCellNum = n;
+        facetIncrsView[*crossingId] = ccc;
+        crossingParentIdsView[*crossingId] = n;
         ++(*crossingId);
       }
     };
 
 #if defined(AXOM_USE_RAJA)
     /*
-      The m_crossings filling loop isn't data-parallel and shouldn't
-      be parallelized.  This contrived RAJA::forall forces it to run
-      sequentially.
+      loopBody isn't data-parallel and shouldn't be parallelized.
+      This contrived RAJA::forall forces it to run sequentially.
     */
     RAJA::forall<SequentialLoopPolicy>(
       RAJA::RangeSegment(0, 1),
@@ -346,14 +426,9 @@ public:
 
     axom::deallocate(crossingId);
 
-    axom::Array<axom::IndexType, 1, MemorySpace> prefixSum(m_crossingCount,
-                                                           m_crossingCount);
-    const auto prefixSumView = prefixSum.view();
+    // axom::Array<axom::IndexType, 1, MemorySpace> prefixSum(m_crossingCount, m_crossingCount);
+    const auto firstFacetIdsView = m_firstFacetIds.view();
 
-    auto copyFirstSurfaceCellId = AXOM_LAMBDA(axom::IndexType n)
-    {
-      crossingsView[n].firstSurfaceCellId = prefixSumView[n];
-    };
 #if defined(AXOM_USE_RAJA)
       // Intel oneAPI compiler segfaults with OpenMP RAJA scan
   #ifdef __INTEL_LLVM_COMPILER
@@ -362,45 +437,81 @@ public:
   #else
     using ScanPolicy = typename axom::execution_space<ExecSpace>::loop_policy;
   #endif
-    RAJA::exclusive_scan<ScanPolicy>(
-      RAJA::make_span(addCellsView.data(), m_crossingCount),
-      RAJA::make_span(prefixSumView.data(), m_crossingCount),
+    RAJA::inclusive_scan<ScanPolicy>(
+      RAJA::make_span(facetIncrsView.data(), m_crossingCount),
+      RAJA::make_span(firstFacetIdsView.data() + 1, m_crossingCount),
       RAJA::operators::plus<axom::IndexType> {});
-    RAJA::forall<LoopPolicy>(RAJA::RangeSegment(0, m_crossingCount),
-                             copyFirstSurfaceCellId);
 #else
     if(m_crossingCount > 0)
     {
-      prefixSumView[0] = 0;
-      for(axom::IndexType i = 1; i < m_crossingCount; ++i)
+      firstFacetIdsView[0] = 0;
+      for(axom::IndexType i = 1; i < 1 + m_crossingCount; ++i)
       {
-        prefixSumView[i] = prefixSumView[i - 1] + addCellsView[i - 1];
-      }
-      for(axom::IndexType i = 0; i < m_crossingCount; ++i)
-      {
-        copyFirstSurfaceCellId(i);
+        firstFacetIdsView[i] = firstFacetIdsView[i - 1] + facetIncrsView[i - 1];
       }
     }
 #endif
+    axom::copy(&m_facetCount,
+               m_firstFacetIds.data() + m_firstFacetIds.size() - 1,
+               sizeof(axom::IndexType));
+    m_firstFacetIds.resize(m_crossingCount);
+  }
 
-    // Data from the last crossing tells us how many contour cells there are.
-    if(m_crossings.empty())
+  void computeContour() override
+  {
+    const auto facetIncrsView = m_facetIncrs.view();
+    const auto firstFacetIdsView = m_firstFacetIds.view();
+    const auto crossingParentIdsView = m_crossingParentIds.view();
+    const auto caseIdsView = m_caseIds.view();
+
+    // Internal contour mesh data to populate
+    axom::ArrayView<axom::IndexType, 2> facetNodeIdsView = m_facetNodeIds;
+    axom::ArrayView<double, 2> facetNodeCoordsView = m_facetNodeCoords;
+    axom::ArrayView<axom::IndexType> facetParentIdsView = m_facetParentIds;
+    const axom::IndexType facetIndexOffset = m_facetIndexOffset;
+
+    ComputeContour_Util ccu(m_contourVal,
+                            m_caseIds.strides(),
+                            m_fcnView,
+                            m_coordsViews);
+
+    auto gen_for_parent_cell = AXOM_LAMBDA(axom::IndexType crossingId)
     {
-      m_contourCellCount = 0;
-    }
-    else
-    {
-      CrossingInfo back;
-      axom::copy(&back,
-                 m_crossings.data() + m_crossings.size() - 1,
-                 sizeof(CrossingInfo));
-      m_contourCellCount =
-        back.firstSurfaceCellId + num_contour_cells(back.caseNum);
-    }
+      auto parentCellId = crossingParentIdsView[crossingId];
+      auto caseId = caseIdsView.flatIndex(parentCellId);
+      Point cornerCoords[CELL_CORNER_COUNT];
+      double cornerValues[CELL_CORNER_COUNT];
+      ccu.get_corner_coords_and_values(parentCellId, cornerCoords, cornerValues);
+
+      auto additionalFacets = facetIncrsView[crossingId];
+      auto firstFacetId = facetIndexOffset + firstFacetIdsView[crossingId];
+
+      for(axom::IndexType fId = 0; fId < additionalFacets; ++fId)
+      {
+        axom::IndexType newFacetId = firstFacetId + fId;
+        axom::IndexType firstCornerId = newFacetId * DIM;
+
+        facetParentIdsView[newFacetId] = parentCellId;
+
+        for(axom::IndexType d = 0; d < DIM; ++d)
+        {
+          axom::IndexType newCornerId = firstCornerId + d;
+          facetNodeIdsView[newFacetId][d] = newCornerId;
+
+          int edge = cases_table(caseId, fId * DIM + d);
+          ccu.linear_interp(edge,
+                            cornerCoords,
+                            cornerValues,
+                            &facetNodeCoordsView(newCornerId, 0));
+        }
+      }
+    };
+
+    axom::for_all<ExecSpace>(0, m_crossingCount, gen_for_parent_cell);
   }
 
   /*!
-    @brief Implementation used by MarchingCubesHybridParallel::computeContour().
+    @brief Implementation used by MarchingCubesImpl::computeContour().
     containing just the objects needed for that part, to be made available
     on devices.
   */
@@ -426,14 +537,14 @@ public:
 
     template <int TDIM = DIM>
     AXOM_HOST_DEVICE typename std::enable_if<TDIM == 2>::type
-    get_corner_coords_and_values(IndexType cellNum,
+    get_corner_coords_and_values(IndexType parentCellId,
                                  Point cornerCoords[],
                                  double cornerValues[]) const
     {
       const auto& x = coordsViews[0];
       const auto& y = coordsViews[1];
 
-      const auto c = indexer.toMultiIndex(cellNum);
+      const auto c = indexer.toMultiIndex(parentCellId);
       const auto& i = c[0];
       const auto& j = c[1];
 
@@ -451,7 +562,7 @@ public:
     }
     template <int TDIM = DIM>
     AXOM_HOST_DEVICE typename std::enable_if<TDIM == 3>::type
-    get_corner_coords_and_values(IndexType cellNum,
+    get_corner_coords_and_values(IndexType parentCellId,
                                  Point cornerCoords[],
                                  double cornerValues[]) const
     {
@@ -459,7 +570,7 @@ public:
       const auto& y = coordsViews[1];
       const auto& z = coordsViews[2];
 
-      const auto c = indexer.toMultiIndex(cellNum);
+      const auto c = indexer.toMultiIndex(parentCellId);
       const auto& i = c[0];
       const auto& j = c[1];
       const auto& k = c[2];
@@ -491,7 +602,7 @@ public:
       int edgeIdx,
       const Point cornerCoords[4],
       const double nodeValues[4],
-      Point& crossingPt) const
+      double* /* Point& */ crossingPt) const
     {
       // STEP 0: get the edge node indices
       // 2 nodes define the edge.  n1 and n2 are the indices of
@@ -512,13 +623,13 @@ public:
       if(axom::utilities::isNearlyEqual(contourVal, f1) ||
          axom::utilities::isNearlyEqual(f1, f2))
       {
-        crossingPt = p1;
+        crossingPt[0] = p1[0]; crossingPt[1] = p1[1]; // crossingPt = p1;
         return;
       }
 
       if(axom::utilities::isNearlyEqual(contourVal, f2))
       {
-        crossingPt = p2;
+        crossingPt[0] = p2[0]; crossingPt[1] = p2[1]; // crossingPt = p2;
         return;
       }
 
@@ -538,7 +649,7 @@ public:
       int edgeIdx,
       const Point cornerCoords[8],
       const double nodeValues[8],
-      Point& crossingPt) const
+      double* /* Point& */ crossingPt) const
     {
       // STEP 0: get the edge node indices
       // 2 nodes define the edge.  n1 and n2 are the indices of
@@ -565,13 +676,13 @@ public:
       if(axom::utilities::isNearlyEqual(contourVal, f1) ||
          axom::utilities::isNearlyEqual(f1, f2))
       {
-        crossingPt = p1;
+        crossingPt[0] = p1[0]; crossingPt[1] = p1[1]; crossingPt[2] = p1[2]; // crossingPt = p1;
         return;
       }
 
       if(axom::utilities::isNearlyEqual(contourVal, f2))
       {
-        crossingPt = p2;
+        crossingPt[0] = p2[0]; crossingPt[1] = p2[1]; crossingPt[2] = p2[2]; // crossingPt = p2;
         return;
       }
 
@@ -585,123 +696,6 @@ public:
       }
     }
   };  // ComputeContour_Util
-
-  void computeContour() override
-  {
-#if 1
-    auto crossingsView = m_crossings.view();
-
-    // Internal contour mesh data to populate
-    axom::ArrayView<double, 2> facetNodeIdsView = m_facetNodeIds;
-    axom::ArrayView<axom::IndexType, 2> facetNodeCoordsView = m_facetNodeCoords;
-    axom::ArrayView<axom::IndexType> facetParentIdsView = m_facetParentIds;
-    const axom::IndexType facetIndexOffset = m_facetIndexOffset;
-
-    ComputeContour_Util ccu(m_contourVal,
-                            m_caseIds.strides(),
-                            m_fcnView,
-                            m_coordsViews);
-
-    auto loopBody = AXOM_LAMBDA(axom::IndexType iCrossing)
-    {
-      const auto& crossingInfo = crossingsView[iCrossing];
-      const IndexType crossingCellCount = num_contour_cells(crossingInfo.caseNum);
-      SLIC_ASSERT(crossingCellCount > 0);
-
-      // Parent cell data for interpolating new node coordinates.
-      Point cornerCoords[CELL_CORNER_COUNT];
-      double cornerValues[CELL_CORNER_COUNT];
-      ccu.get_corner_coords_and_values(crossingInfo.parentCellNum,
-                                       cornerCoords,
-                                       cornerValues);
-
-      /*
-        Create the new cell and its DIM nodes.  New nodes are on
-        parent cell edges where the edge intersects the isocontour.
-        linear_interp for the exact coordinates.
-      */
-      for(int iCell = 0; iCell < crossingCellCount; ++iCell)
-      {
-        IndexType newFacetId = facetIndexOffset + crossingInfo.firstSurfaceCellId + iCell;
-        facetParentIdsView[newFacetId] = crossingInfo.parentCellNum;
-        for(int d = 0; d < DIM; ++d)
-        {
-          IndexType newNodeId = newFacetId * DIM + d;
-          axom::StackArray<axom::IndexType, 2> idx{newFacetId, d};
-          facetNodeIdsView[idx] = newNodeId;
-
-          const int edge = cases_table(crossingInfo.caseNum, iCell * DIM + d);
-          ccu.linear_interp(edge,
-                            cornerCoords,
-                            cornerValues,
-                            facetNodeCoordsView[newNodeId]);
-        }
-      }
-    };
-    axom::for_all<ExecSpace>(0, m_crossingCount, loopBody);
-#else
-    auto crossingsView = m_crossings.view();
-
-    /*
-      Reserve contour mesh data space so we can add data without
-      reallocation.
-    */
-    const axom::IndexType contourNodeCount = DIM * m_contourCellCount;
-    m_facetNodeCoords.resize(contourNodeCount);
-    m_facetNodeIds.resize(m_contourCellCount);
-    m_facetParentIds.resize(m_contourCellCount);
-
-    auto nodeCoordsView = m_facetNodeCoords.view();
-    auto cellCornersView = m_facetNodeIds.view();
-    auto cellParentsView = m_facetParentIds.view();
-
-    ComputeContour_Util ccu(m_contourVal,
-                            m_caseIds.strides(),
-                            m_fcnView,
-                            m_coordsViews);
-
-    auto loopBody = AXOM_LAMBDA(axom::IndexType iCrossing)
-    {
-      const auto& crossingInfo = crossingsView[iCrossing];
-      const IndexType crossingCellCount = num_contour_cells(crossingInfo.caseNum);
-      SLIC_ASSERT(crossingCellCount > 0);
-
-      // Parent cell data for interpolating new node coordinates.
-      Point cornerCoords[CELL_CORNER_COUNT];
-      double cornerValues[CELL_CORNER_COUNT];
-      ccu.get_corner_coords_and_values(crossingInfo.parentCellNum,
-                                       cornerCoords,
-                                       cornerValues);
-
-      /*
-        Create the new cell and its DIM nodes.  New nodes are on
-        parent cell edges where the edge intersects the isocontour.
-        linear_interp for the exact coordinates.
-
-        TODO: The varying crossingCellCount value may inhibit device
-        performance.  Try grouping m_crossings items that have the
-        same values for crossingCellCount.
-      */
-      for(int iCell = 0; iCell < crossingCellCount; ++iCell)
-      {
-        IndexType contourCellId = crossingInfo.firstSurfaceCellId + iCell;
-        cellParentsView[contourCellId] = crossingInfo.parentCellNum;
-        for(int d = 0; d < DIM; ++d)
-        {
-          IndexType contourNodeId = contourCellId * DIM + d;
-          cellCornersView[contourCellId][d] = contourNodeId;
-
-          const int edge = cases_table(crossingInfo.caseNum, iCell * DIM + d);
-          ccu.linear_interp(edge,
-                            cornerCoords,
-                            cornerValues,
-                            nodeCoordsView[contourNodeId]);
-        }
-      }
-    };
-    axom::for_all<ExecSpace>(0, m_crossingCount, loopBody);
-#endif
-  }
 
   // These 4 functions provide access to the look-up table
   // whether on host or device.  Is there a more elegant way
@@ -751,106 +745,6 @@ public:
     return cases3D[iCase][iEdge];
   }
 
-#if 0
-  /*!
-    @brief Output contour mesh to a mint::UnstructuredMesh object.
-  */
-  void populateContourMesh(
-    axom::mint::UnstructuredMesh<axom::mint::SINGLE_SHAPE>& mesh,
-    const std::string& cellIdField) const override
-  {
-    auto internalAllocatorID = axom::execution_space<ExecSpace>::allocatorID();
-    auto hostAllocatorID = axom::execution_space<axom::SEQ_EXEC>::allocatorID();
-
-    /*
-      mint uses host memory.  If internal memory is on the host, use
-      it.  Otherwise, make a temporary copy of it on the host.
-    */
-    if(internalAllocatorID == hostAllocatorID)
-    {
-      populateContourMesh(mesh,
-                          cellIdField,
-                          m_facetNodeCoords,
-                          m_facetNodeIds,
-                          m_facetParentIds);
-    }
-    else
-    {
-      axom::Array<Point, 1, MemorySpace::Dynamic> facetNodeCoords(
-        m_facetNodeCoords,
-        hostAllocatorID);
-      axom::Array<MIdx, 1, MemorySpace::Dynamic> facetNodeIds(
-        m_facetNodeIds,
-        hostAllocatorID);
-      axom::Array<IndexType, 1, MemorySpace::Dynamic> facetParentIds(
-        m_facetParentIds,
-        hostAllocatorID);
-
-      populateContourMesh(mesh,
-                          cellIdField,
-                          facetNodeCoords,
-                          facetNodeIds,
-                          facetParentIds);
-    }
-  }
-
-  //!@brief Output contour mesh to a mint::UnstructuredMesh object.
-  void populateContourMesh(
-    axom::mint::UnstructuredMesh<axom::mint::SINGLE_SHAPE>& mesh,
-    const std::string& cellIdField,
-    const axom::Array<Point, 1, MemorySpace::Dynamic> facetNodeCoords,
-    const axom::Array<MIdx, 1, MemorySpace::Dynamic> facetNodeIds,
-    const axom::Array<IndexType, 1, MemorySpace::Dynamic> facetParentIds) const
-  {
-    // AXOM_PERF_MARK_FUNCTION("MarchingCubesHybridParallel::populateContourMesh");
-    if(!cellIdField.empty() &&
-       !mesh.hasField(cellIdField, axom::mint::CELL_CENTERED))
-    {
-      mesh.createField<axom::IndexType>(cellIdField,
-                                        axom::mint::CELL_CENTERED,
-                                        DIM);
-    }
-
-    const axom::IndexType addedCellCount = facetNodeIds.size();
-    const axom::IndexType addedNodeCount = facetNodeCoords.size();
-    if(addedCellCount != 0)
-    {
-      const axom::IndexType priorCellCount = mesh.getNumberOfCells();
-      const axom::IndexType priorNodeCount = mesh.getNumberOfNodes();
-      mesh.reserveNodes(priorNodeCount + addedNodeCount);
-      mesh.reserveCells(priorCellCount + addedCellCount);
-
-      mesh.appendNodes((double*)facetNodeCoords.data(),
-                       facetNodeCoords.size());
-      for(int n = 0; n < addedCellCount; ++n)
-      {
-        MIdx cornerIds = facetNodeIds[n];
-        // Bump corner indices by priorNodeCount to avoid indices
-        // used by other parents domains.
-        for(int d = 0; d < DIM; ++d)
-        {
-          cornerIds[d] += priorNodeCount;
-        }
-        mesh.appendCell(cornerIds);
-      }
-      axom::IndexType numComponents = -1;
-      axom::IndexType* cellIdPtr =
-        mesh.getFieldPtr<axom::IndexType>(cellIdField,
-                                          axom::mint::CELL_CENTERED,
-                                          numComponents);
-      SLIC_ASSERT(numComponents == DIM);
-      axom::ArrayView<axom::StackArray<axom::IndexType, DIM>> cellIdView(
-        (axom::StackArray<axom::IndexType, DIM>*)cellIdPtr,
-        priorCellCount + addedCellCount);
-      axom::ArrayIndexer<axom::IndexType, DIM> si(m_caseIds.shape(), 'c');
-      for(axom::IndexType i = 0; i < addedCellCount; ++i)
-      {
-        cellIdView[priorCellCount + i] = si.toMultiIndex(facetParentIds[i]);
-      }
-    }
-  }
-#endif
-
   //!@brief Compute the case index into cases2D or cases3D.
   AXOM_HOST_DEVICE inline int compute_crossing_case(const double* f) const
   {
@@ -866,40 +760,11 @@ public:
     return index;
   }
 
-#if 0
-  //!@brief Clear data so you can rerun with a different contour value.
-  void clear()
-  {
-    m_facetNodeCoords.clear();
-    m_facetNodeIds.clear();
-    m_facetParentIds.clear();
-    m_crossingCount = 0;
-    m_contourCellCount = 0;
-  }
-#endif
-
   /*!
     @brief Constructor.
   */
-  MarchingCubesHybridParallel()
-    : m_crossings(0, 0)
+  MarchingCubesImpl()
   { }
-
-  /*!
-    @brief Info for a parent cell intersecting the contour surface.
-  */
-  struct CrossingInfo
-  {
-    CrossingInfo() { }
-    CrossingInfo(axom::IndexType parentCellNum_, std::uint16_t caseNum_)
-      : parentCellNum(parentCellNum_)
-      , caseNum(caseNum_)
-      , firstSurfaceCellId(std::numeric_limits<axom::IndexType>::max())
-    { }
-    axom::IndexType parentCellNum;  //!< @brief Flat index of parent cell in m_caseIds.
-    std::uint16_t caseNum;          //!< @brief Index in cases2D or cases3D
-    axom::IndexType firstSurfaceCellId;  //!< @brief First index for generated cells.
-  };
 
 private:
   axom::quest::MeshViewUtil<DIM, MemorySpace> m_mvu;
@@ -916,18 +781,21 @@ private:
   //!@brief Crossing case for each computational mesh cell.
   axom::Array<std::uint16_t, DIM, MemorySpace> m_caseIds;
 
-  //!@brief Info on every parent cell that crosses the contour surface.
-  axom::Array<CrossingInfo, 1, MemorySpace> m_crossings;
-
   //!@brief Number of parent cells crossing the contour surface.
   axom::IndexType m_crossingCount = 0;
 
-  //!@brief Number of contour surface cells from crossings.
-  axom::IndexType m_contourCellCount = 0;
-  axom::IndexType getContourCellCount() const override
-  {
-    return m_contourCellCount;
-  }
+  //!@brief Number of contour surface cells from all crossings.
+  axom::IndexType m_facetCount = 0;
+  axom::IndexType getContourCellCount() const override { return m_facetCount; }
+
+  //!@brief Number of surface mesh facets added by each crossing.
+  axom::Array<FacetIdType, 1, MemorySpace> m_facetIncrs;
+
+  //!@brief Parent cell id (flat index into m_caseIds) for each crossing.
+  axom::Array<axom::IndexType, 1, MemorySpace> m_crossingParentIds;
+
+  //!@brief First index of facets for each crossing.
+  axom::Array<axom::IndexType, 1, MemorySpace> m_firstFacetIds;
 
   //!@brief Number of corners (nodes) on each parent cell.
   static constexpr std::uint8_t CELL_CORNER_COUNT = (DIM == 3) ? 8 : 4;
@@ -936,11 +804,11 @@ private:
 };
 
 /*!
-  @brief Allocate a MarchingCubesHybridParallel object, template-specialized
+  @brief Allocate a MarchingCubesImpl object, template-specialized
   for caller-specified runtime policy and physical dimension.
 */
 static std::unique_ptr<axom::quest::MarchingCubesSingleDomain::ImplBase>
-newMarchingCubesHybridParallel(MarchingCubes::RuntimePolicy runtimePolicy, int dim)
+newMarchingCubesImpl(MarchingCubes::RuntimePolicy runtimePolicy, int dim)
 {
   using ImplBase = axom::quest::MarchingCubesSingleDomain::ImplBase;
 
@@ -948,40 +816,36 @@ newMarchingCubesHybridParallel(MarchingCubes::RuntimePolicy runtimePolicy, int d
   std::unique_ptr<ImplBase> impl;
   if(runtimePolicy == MarchingCubes::RuntimePolicy::seq)
   {
-    impl = dim == 2
-      ? std::unique_ptr<ImplBase>(
-          new MarchingCubesHybridParallel<2, axom::SEQ_EXEC, axom::SEQ_EXEC>)
-      : std::unique_ptr<ImplBase>(
-          new MarchingCubesHybridParallel<3, axom::SEQ_EXEC, axom::SEQ_EXEC>);
+    impl = dim == 2 ? std::unique_ptr<ImplBase>(
+                        new MarchingCubesImpl<2, axom::SEQ_EXEC, axom::SEQ_EXEC>)
+                    : std::unique_ptr<ImplBase>(
+                        new MarchingCubesImpl<3, axom::SEQ_EXEC, axom::SEQ_EXEC>);
   }
 #ifdef AXOM_RUNTIME_POLICY_USE_OPENMP
   else if(runtimePolicy == MarchingCubes::RuntimePolicy::omp)
   {
-    impl = dim == 2
-      ? std::unique_ptr<ImplBase>(
-          new MarchingCubesHybridParallel<2, axom::OMP_EXEC, axom::SEQ_EXEC>)
-      : std::unique_ptr<ImplBase>(
-          new MarchingCubesHybridParallel<3, axom::OMP_EXEC, axom::SEQ_EXEC>);
+    impl = dim == 2 ? std::unique_ptr<ImplBase>(
+                        new MarchingCubesImpl<2, axom::OMP_EXEC, axom::SEQ_EXEC>)
+                    : std::unique_ptr<ImplBase>(
+                        new MarchingCubesImpl<3, axom::OMP_EXEC, axom::SEQ_EXEC>);
   }
 #endif
 #ifdef AXOM_RUNTIME_POLICY_USE_CUDA
   else if(runtimePolicy == MarchingCubes::RuntimePolicy::cuda)
   {
-    impl = dim == 2
-      ? std::unique_ptr<ImplBase>(
-          new MarchingCubesHybridParallel<2, axom::CUDA_EXEC<256>, axom::CUDA_EXEC<1>>)
-      : std::unique_ptr<ImplBase>(
-          new MarchingCubesHybridParallel<3, axom::CUDA_EXEC<256>, axom::CUDA_EXEC<1>>);
+    impl = dim == 2 ? std::unique_ptr<ImplBase>(
+                        new MarchingCubesImpl<2, axom::CUDA_EXEC<256>, axom::CUDA_EXEC<1>>)
+                    : std::unique_ptr<ImplBase>(
+                        new MarchingCubesImpl<3, axom::CUDA_EXEC<256>, axom::CUDA_EXEC<1>>);
   }
 #endif
 #ifdef AXOM_RUNTIME_POLICY_USE_HIP
   else if(runtimePolicy == MarchingCubes::RuntimePolicy::hip)
   {
-    impl = dim == 2
-      ? std::unique_ptr<ImplBase>(
-          new MarchingCubesHybridParallel<2, axom::HIP_EXEC<256>, axom::HIP_EXEC<1>>)
-      : std::unique_ptr<ImplBase>(
-          new MarchingCubesHybridParallel<3, axom::HIP_EXEC<256>, axom::HIP_EXEC<1>>);
+    impl = dim == 2 ? std::unique_ptr<ImplBase>(
+                        new MarchingCubesImpl<2, axom::HIP_EXEC<256>, axom::HIP_EXEC<1>>)
+                    : std::unique_ptr<ImplBase>(
+                        new MarchingCubesImpl<3, axom::HIP_EXEC<256>, axom::HIP_EXEC<1>>);
   }
 #endif
   else
